+++ src/main/java/com/somemore/volunteer/service/VolunteerQueryService.java
@@ -2,7 +2,6 @@ package com.somemore.volunteer.service;
 
 import static com.somemore.global.exception.ExceptionMessage.NOT_EXISTS_VOLUNTEER;
 
-import com.somemore.facade.validator.VolunteerDetailAccessValidator;
 import com.somemore.global.exception.BadRequestException;
 import com.somemore.volunteer.domain.Volunteer;
 import com.somemore.volunteer.domain.VolunteerDetail;
@@ -13,6 +12,7 @@ import com.somemore.volunteer.repository.VolunteerRepository;
 import com.somemore.volunteer.repository.mapper.VolunteerOverviewForRankingByHours;
 import com.somemore.volunteer.repository.mapper.VolunteerSimpleInfo;
 import com.somemore.volunteer.usecase.VolunteerQueryUseCase;
+import com.somemore.volunteer.validator.VolunteerDetailAccessValidator;
 import java.util.List;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
