+++ src/main/java/com/somemore/volunteerapply/controller/VolunteerApplyQueryApiController.java
@@ -3,7 +3,7 @@ package com.somemore.volunteerapply.controller;
 import static org.springframework.data.domain.Sort.Direction.DESC;
 
 import com.somemore.global.auth.annotation.CurrentUser;
-import com.somemore.facade.volunteerapply.VolunteerApplyQueryFacadeUseCase;
+import com.somemore.volunteerapply.usecase.VolunteerApplyQueryFacadeUseCase;
 import com.somemore.global.common.response.ApiResponse;
 import com.somemore.global.exception.BadRequestException;
 import com.somemore.volunteerapply.domain.ApplyStatus;
