+++ src/test/java/com/somemore/volunteerapply/controller/CenterVolunteerApplyCommandApiControllerTest.java
@@ -11,7 +11,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.somemore.support.ControllerTestSupport;
 import com.somemore.support.annotation.WithMockCustomUser;
-import com.somemore.facade.volunteerapply.SettleVolunteerApplyFacadeUseCase;
+import com.somemore.volunteerapply.usecase.SettleVolunteerApplyFacadeUseCase;
 import com.somemore.volunteerapply.dto.request.VolunteerApplySettleRequestDto;
 import com.somemore.volunteerapply.usecase.ApproveVolunteerApplyUseCase;
 import com.somemore.volunteerapply.usecase.RejectVolunteerApplyUseCase;
