+++ src/test/java/com/somemore/volunteerapply/controller/VolunteerApplyQueryApiControllerTest.java
@@ -15,7 +15,7 @@ import com.somemore.volunteerapply.dto.response.VolunteerApplyRecruitInfoRespons
 import com.somemore.volunteerapply.dto.response.VolunteerApplyResponseDto;
 import com.somemore.volunteerapply.dto.response.VolunteerApplySummaryResponseDto;
 import com.somemore.volunteerapply.dto.response.VolunteerApplyVolunteerInfoResponseDto;
-import com.somemore.facade.volunteerapply.VolunteerApplyQueryFacadeUseCase;
+import com.somemore.volunteerapply.usecase.VolunteerApplyQueryFacadeUseCase;
 import com.somemore.volunteerapply.usecase.VolunteerApplyQueryUseCase;
 import java.util.Collections;
 import java.util.UUID;
